chore: refactor getAgeInMonth function into a separate file
